move sockets out of server.js to controller
