+++ server.js
@@ -5,8 +5,6 @@ const cors = require('cors');
 const passport = require('passport');
 require("dotenv");
 
-const { addUser, removeUser, getUser, getUsersInGame, addResponses, getResponses } = require('./controllers/userController');
-const { roomTimer, activeTimer } = require("./controllers/roomTimer");
 const db = require("./models");
 
 const path = require("path");
@@ -37,63 +35,7 @@ app.get("/*", (req, res) => {
   res.sendFile(path.join(__dirname, 'client/build', 'index.html'));
 });
 
-// Set up socket handlers
-io.on('connect', (socket) => {
-
-  socket.on('join', ({ game, name, icon, color }, callback) => {
-    console.log(`Socket id on join: ${socket.id}`);
-
-    const { error, user } = addUser({ id: socket.id, game, name, icon, color });
-
-    if(error) return callback(error);
-
-    socket.join(user.game);
-
-    //socket.emit('arrival', { user });
-    //socket.broadcast.to(user.game).emit('arrival', { user });
-
-    io.to(user.game).emit('gameData', { game: user.game, users: getUsersInGame(user.game)});
-
-    callback();
-  })
-
-  socket.on("allHere", ({ game }, callback) => {
-    console.log(`Socket id on allHere: ${socket.id}`);
-    const user = getUser(socket.id);
-
-    io.to(user.game).emit("startGame", { game: user.game, users: getUsersInGame(user.game)});
-  })
-
-  socket.on('response', async ({ game }, callback) => {
-    //get the user from the server based on the socket id
-    const user = await getUser(socket.id);
-
-    io.to(user.game).emit('respData', {game: user.game});
-  })
-
-  socket.on("startquestion", async ({ game }, callback) => {
-    const user = await getUser(socket.id);
-    console.log(`Socket id on startquestion: ${socket.id}`);
-    //create a new roomTimer if there isn't already one active
-    roomTimer(user.game, "question", io);
-  })
-
-  socket.on('scoringComplete', async ({ game }, callback) => {
-    const user = await getUser(socket.id);
-    console.log(`Scoring is complete - socketid: ${socket.id}`)
-    //add a pause so they can see their scores
-    roomTimer(user.game, "pause", io);
-  })
-
-  socket.on('disconnect', () => {
-    console.log("socket disconnect request received");
-    const user = removeUser(socket.id);
-
-    if(user){
-      io.to(user.game).emit('departure', { user: user.name, text: `${user.name} has left`})
-    }
-  })
-})
+require("./controllers/sockets.js")(io);
 
 //sync the db with sequelize
 db.sequelize.sync().then(function () {
